+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/TabbedApplicationLayout.java
@@ -112,6 +112,7 @@ public class TabbedApplicationLayout implements EntityApplicationPanel.Applicati
 	private void addTab(EntityPanel entityPanel) {
 		tabbedPane.addTab(entityPanel.caption().get(), entityPanel);
 		tabbedPane.setToolTipTextAt(tabbedPane.getTabCount() - 1, entityPanel.description().get());
+		tabbedPane.setIconAt(tabbedPane.getTabCount() - 1, entityPanel.icon().get());
 	}
 
 	private final class InitializeSelectedPanelListener implements ChangeListener {